Remove user-specific features to concentrate on MVP
This commit removes user-specific classes (RestController, SecurityConfiguration, User, UserAccessRepo, UserServiceImpl) from the project. This is done to focus primarily on the Minimum Viable Product (MVP) features. With this change, user-related features and authorizations will be handled later on. The pom.xml and MessageAccessRepo are updated, and Message.java has been significantly changed. New jsp-pages are added, and AppConfig and MappingController classes in place now. The user-related features removed and changes are made to ensure the stable functioning of other project features. It's a major shift in the project, but it is necessary to ensure successful completion of the MVP before extending user-specific features.