Refactor Controller, Add Services, Load Test and Scheduling features
Refactored MappingController to use MessageService instead of accessing repository directly as per the standard layered architecture. MessageService.java is created to handle business logic and database operations. The new class LoadTest.java was created to provide a utility for load testing. ScheduledOperations.java was added to schedule operations such as deleting messages marked as deleted, making the upkeep of the database automatic and efficient. Also, pom.xml was updated to include a testing library awaitility and the application was enabled to support scheduling. This also includes changes in exception handling, standard CRUD operations and viewResolver properties.