"Refactor checkAuthentication method and related updates"
Refactored the 'checkAuthentication' method in 'CheckIfUserLoggedIn' Aspect to check if the user is already logged in. A custom service 'AuthenticationStatus' service is used now instead of directly using the 'SecurityContextHolder'. This service checks for null pointers before checking authentication making the code more robust Now, it prevents already logged-in users from accessing login or registration page. .