Improve message deletion process and clean up code
Added cache invalidation to the process of deleting messages from the repository in MessageService. This was done to ensure that cache doesn't serve stale data. Also, made changes to Repository to return the list of deleted messages.

Cleaned up code in UserMapping and removed a TODO comment that was no longer relevant.

Removed unnecessary comments from the pom.xml file for better readability. Additionally, changed the project SDK configuration to use Oracle JDK 21 in .idea/misc.xml.

Updated Kafka client library to version 3.6.0 and removed h2 database dependency since it is not used.