Remove last sequence file, refactor sequence handling
The last sequence value for generating short URLs has been refactored to be fetched from database instead of the deleted file "last_sequence.txt". The method getLastGeneratedSequence in ShortURLRepo and ShortURLService now returns an Optional< String > instead of String. If the table is empty, sequence "a" is returned as default. Fetching the last generated amount was also refactored; files are replaced with database queries. The redundant method setLastGeneratedValue was removed.