Refactor error handling and improve user authentication
The error handling process was refactored to allow for more flexibility and improve readability. Model object was used directly instead of accessing through modelAndView for adding exception attributes. In addition, method 'loadUserByUsername' under UserDetailsService now throws 'NoSuchElementException' instead of returning null, upon not finding the user.

Aspect class 'CheckIfUserLoggedIn' was renamed to 'OnlyIfUserNotLoggedIn' to represent the purpose more appropriately, where it blocks the service if the user is already logged in. It's using 'AuthenticationContext' instead of 'AuthenticationStatus' to check the user's authentication.

In application.properties, additional configurations were added to expose all actuator health endpoints. This was done to assist in easier application health monitoring and diagnostics.