Remove caching from UserService and shorten cache expiry
Caching was removed from UserService because later user will be persisted in AuthenticationContext.java. Instead of throwing an error, the method now returns an optional user, allowing the error to be handled upfront and providing more flexibility.

Also, the cache expiry time was reduced from 30 to 10 minutes in AppConfig for consistency in data and to ensure more frequent updates.