Add Dockerfile and docker-compose for app deployment
The Dockerfile and docker-compose.yml files were added to the project. These files are meant to simplify the process of building, packaging and running the app in a docker container. Dockerfile contains the instructions for building the docker image of the application, while docker-compose.yml is for setting up the app and its database service. With these configurations, both development and deployment processes become more consistent and manageable, as well as the app's environment becomes easily reproducible.

Right now container app badly works with database (despite the fact that database container works perfectly), fixing...