Update Docker setup
The Docker file and compose files have been updated to improve the setup for the application. The SPRING_PROFILES_ACTIVE environment variable has been added to the application Docker file to ensure the Docker profile is set when running in a container. The version of the Docker Compose file has been updated, and the POSTGRES_DB and SPRING_DATASOURCE_URL environment variables have been updated for consistency. Now containers will launch properly.