Refactor service imports and add active profiles for testing
In the test files, the imports for 'MessageService' and 'ShortURLService' were updated to reflect changes in their package structure. This was necessary due to a prior refactoring of the service classes into a new 'entityService' package. Additionally, the '@ActiveProfiles' annotation was added to 'MessageMappingTest' to set testing profiles to 'local'.