Replace file usage with DB queries in ShortURL class
The last generated amount was previously retrieved from a file, this has been changed to a DB query. The DB query is handled by a newly created method in the ShortURLGenerationMetadata class, known as getGenerationValue. This change makes the software more robust and efficient by avoiding I/O operations and potential file-related errors.