"Add Spring profile argument to Dockerfile"
A SPRING_ACTIVE_PROFILE argument was added to the Dockerfile and docker-compose.yml files. The purpose of these changes is to allow specifying what Spring profile to use when running the application inside Docker. This way, application can run in environment specified by this profile which can be different from a default one.