Add active Spring profile as build argument in Docker
Included a new ARG parameter in the Dockerfile, "SPRING_ACTIVE_PROFILE", to allow the specification of different Spring active profiles during the Docker image build. The active profile is now passed as an argument during the Docker build process in the docker-compose.yml file. This flexibility is needed to properly segregate configurations based on the environment we're deploying the Dockerized Spring application.