Refactor UserDetailsService to adhere to Spring Security
Removed AuthenticationProvider and related classes from the UserDetailsService. Instead, we are using SpringSecurity's UserDetailsService directly. Removed an unnecessary call to findUserByEmail since we are already assuming the user is checked and is present in the database. This is to reduce redundancy, adhere to best practices and minimize the potential for security flaws.