Refactor UserDTO and simplify UserDetails
Removed unnecessary imports from UserDetails as authentication is being handled elsewhere and these were unused. Added additional constructors and a new property to UserDTO to allow for flexibility when constructing a UserDTO, in particular to accommodate users with existing messages. Removed the login method from RootMapping because authentication is re-structured.