Update application settings, refactor UI and cleanup codebase
This commit contains several changes across the front-end and back-end of the application. On the back-end, we've added and updated settings in the application properties files for test and local profiles to improve system management and configuration. Unnecessary dependencies and code from several classes and tests were removed to tidy the codebase.

On the front-end, both `new_message.html` and `get_message.html` were significantly refactored to incorporate the Bootstrap framework for a cleaner and more responsive UI. Updates included resized text areas, reorganized form groups, and Bootstrap styles and scripts added.