Update Postgres version and fix detached entity issue
This commit addresses two key changes. Firstly, the PostgreSQL database version has been updated to the designated version 16.1 in the docker-compose file to ensure compatibility and includes a restart policy for better availability.

Secondly, a comment has been added to the MessageService.java that marks a TODO to handle a case where a detached entity is being passed to 'persist'. This is to ensure that this issue is addressed in future work due to its potential of causing data persistence problems. Lastly, the pom.xml file is updated to match with the new PostgreSQL version.