Set debug logging level and amend scheduler properties
In this commit, the application's logging level has been switched to debug in the application-docker.properties file. This change helps with troubleshooting and providing more specific details in case of an error.

Moreover, for the ScheduledOperations class, the initial delay for the finalDeleteMessages() method has been set to 3 days. This ensures that the scheduled deletion task doesn't run immediately after the application is started but after a delay of 3 days, thus providing enough time for the data accumulation.